requirements.txt
PyJWT==1.6.1
Django==2.0.4
PyGithub==1.39
Jinja2==2.10
MarkupSafe==1.0
pip==10.0.1
pytz==2018.4
setuptools==39.0.1
wheel==0.31.0